+++ docs/_static/iframe_custom.js
@@ -1,0 +1,6 @@
+document.addEventListener("click", handler, true);
+
+function handler(e) {
+    e.stopPropagation();
+    e.preventDefault();
+}