+++ iommi/table.py
@@ -426,10 +426,10 @@ class Column(Part):
         """
         Parameters with the prefix `filter__` will be passed along downstream to the `Filter` instance if applicable. This can be used to tweak the filtering of a column.
 
-        :param after: Set the order of columns, see the `howto on ordering <https://docs.iommi.rocks/en/latest/cookbook_tables.html#how-do-i-reorder-columns>`_ for an example.
+        :param after: Set the order of columns, see the `howto on ordering <https://docs.iommi.rocks//cookbook_tables.html#how-do-i-reorder-columns>`_ for an example.
         :param attr: What attribute to use, defaults to same as name. Follows django conventions to access properties of properties, so `foo__bar` is equivalent to the python code `foo.bar`. This parameter is based on the filter name of the Column if you use the declarative style of creating tables.
-        :param bulk: Namespace to configure bulk actions. See `howto on bulk editing <https://docs.iommi.rocks/en/latest/cookbook_tables.html#how-do-i-enable-bulk-editing>`_ for an example and more information.
-        :param cell: Customize the cell, see See `howto on rendering <https://docs.iommi.rocks/en/latest/cookbook_tables.html#how-do-i-customize-the-rendering-of-a-cell>`_ and `howto on links <https://docs.iommi.rocks/en/latest/cookbook_tables.html#how-do-i-make-a-link-in-a-cell>`_
+        :param bulk: Namespace to configure bulk actions. See `howto on bulk editing <https://docs.iommi.rocks//cookbook_tables.html#how-do-i-enable-bulk-editing>`_ for an example and more information.
+        :param cell: Customize the cell, see See `howto on rendering <https://docs.iommi.rocks//cookbook_tables.html#how-do-i-customize-the-rendering-of-a-cell>`_ and `howto on links <https://docs.iommi.rocks//cookbook_tables.html#how-do-i-make-a-link-in-a-cell>`_
         :param display_name: the text of the header for this column. By default this is based on the `_name` so normally you won't need to specify it.
         :param url: URL of the header. This should only be used if sorting is off.
         :param include: set this to `False` to hide the column